+++ haiku_rag_slim/haiku/rag/store/repositories/document_item.py
@@ -1,8 +1,9 @@
 import json
 from collections.abc import Mapping, Sequence
 
-from haiku.rag.store.engine import DocumentItemRecord, Store
+from haiku.rag.store.engine import Store
 from haiku.rag.store.models.document_item import DocumentItem
+from haiku.rag.store.schema import DocumentItemRecord
 from haiku.rag.utils import escape_sql_string
 
 # Per-item metadata columns. The payload column ``picture_data`` is fetched
