+++ tests/test_s3_integration.py
@@ -65,7 +65,7 @@ def config():
 
 @pytest.mark.asyncio
 async def test_store_connect_and_create(tmp_path, config):
-    from haiku.rag.store.engine import get_database_stats
+    from haiku.rag.store.info import get_database_stats
 
     async with Store(tmp_path / "unused", config=config, create=True) as store:
         stats = await get_database_stats(store.db)
@@ -81,7 +81,8 @@ async def test_store_vacuum(tmp_path, config):
 
 @pytest.mark.asyncio
 async def test_store_add_document(tmp_path, config):
-    from haiku.rag.store.engine import DocumentRecord, get_database_stats
+    from haiku.rag.store.info import get_database_stats
+    from haiku.rag.store.schema import DocumentRecord
 
     async with Store(tmp_path / "unused", config=config, create=True) as store:
         doc = DocumentRecord(content="The quick brown fox jumps over the lazy dog.")
